App moved to components folder!
